// Kerekparbolt/WebUI/src/services/client/bicycleService.ts
export interface BicycleEntity {
    Ar?: number | undefined;
    Cikkszam?: string | undefined;
    FelniAtmeroID?: number | undefined;
    Id?: number | undefined;
    MarkaID?: number | undefined;
    TipusID?: number | undefined;
    URL?: string | undefined;
    ValtoTipus?: number | undefined;
    VazmeretID?: number | undefined;
}