// Kerekparbolt/WebUI/src/services/client/brandService.ts
export interface BrandEntity {
    Id: number | undefined;
    Name: string | undefined;
    id: number | undefined;
    name: string | undefined;
}