// Kerekparbolt/WebUI/src/services/client/shifterService.ts
export interface ShifterEntity {
    Id: number | undefined;
    Name: string | undefined;
    id: number | undefined;
    name: string | undefined;
}